requirements.txt
timm==1.0.15
pywaggle[all]==0.56.0
huggingface_hub==0.33.1
ultralytics>=8.3.70
torch>=2.0.0
transformers==4.52.4
pillow==10.3.0
numpy==1.24.4
opencv-python==4.11.0.86
requests==2.32.3
open_clip_torch>=2.20.0
